Search v1 Working
Added a mui autocomplete component to the client which uses the api to make a aggregation search request on the mongodb atlas autocomplete search index.

Requires users to setup the index in atlas.

Still need to add handling for not including search, UI/UX aspects, and the functionality for clicking items.