Added New Control & Fixed List Name/Title Issue
Added a new .env variable REACT_APP_Search. Which should be true or false. When true we display the search bar. When false or missing we dont.

Added handling for list display items to show the searchValue in the left portion. Now we dont have the requirement to have a document field called name or title.