Guide & Context Updates
Added context to useFetchFilters and supplied it at the app level. Since it was the logical place to check if we have any visible filters. The value we generate is a boolean called hasVisibleFilters.

The useGuide custom hook can utilize the hasVisibleFilters value to inform our explainerSection component on what to show users.

I intend to add functionality to add more guide elements. Though it may be logical to refactor the code before then and clean up the file structure.