Implementing Redux For Filters
Almost done implementing redux for filters. Which handles the state for when we grab filters from the db. This resulted in a lot cleaner code.

Only stopping since its late and Ill have to rewrite the filter editors display section which consists of old code and weird handling of data.