Filters Redux Implemented
Now whenever a user interacts with the selection of filters it is properly handled by Redux. The UX is identical to before.

This commit handled the cleaning up of rendering the filter editors, which were pretty old code. Was able to remove componentTyper as it was an old way of rendering the components dynamically.