Merged FIlter Slices
Merged the filters and filterSearch slice into filterSearch and moved filters helpers into filterSearchs helpers. This made sense as they used eachothers state often.

I also fixed a few issues with filterSearch and how it worked with the UI, there is now more user feedback as to what is happening.

Fixed an issue with previous search not properly keeping track of the changes to filter parameters.

I also removed useless code and files.